Correciones DTO,Helpers, SQLs y Entitities

+++ FinK/src/main/java/co/edu/uco/fink/entity/SuministroRequeridoEntity.java
@@ -62,7 +62,7 @@ public class SuministroRequeridoEntity {
         this.suministro = ObjectHelper.getObjectHelper().getDefault(suministro, TipoSuministroEntity.build());
     }
 
-    public final float getCantidad() {
+    public final int getCantidad() {
         return cantidad;
     }
